Remove imxrt-dma features
There's no need to statically switch the size of DMA channel arrays.
The number 32 always works, since the register block is consistently
sized across all i.MX RT variants. The only thing that differs is
which index you're permitted to use. We can require the user to do
that without implementing any static guarantees.
